requirements.txt
# Web Framework
fastapi==0.109.0              # Modern async web framework for webhook handling
uvicorn[standard]==0.27.0     # ASGI server to run FastAPI

# Telegram Bot
python-telegram-bot==21.0     # Official Telegram Bot API wrapper

# Google Cloud
google-cloud-firestore==2.14.0     # Firestore database client
google-cloud-aiplatform==1.42.0    # Vertex AI for Gemini (Phase 2)
google-cloud-secret-manager==2.17.0  # Fetch secrets from GCP

# Vertex AI SDK for Gemini (Phase 2)
google-generativeai>=0.3.0        # Gemini API client

# Phase 2: We build the multi-agent system directly (no LangChain needed)
# This avoids dependency conflicts and gives us more control

# Data Validation
pydantic>=2.10.0              # Type-safe data models (Python 3.13 compatible)
pydantic-settings>=2.7.0      # Environment variable validation

# Utilities
python-dotenv==1.0.0          # Load .env file for local development
httpx~=0.27.0                 # Async HTTP client (compatible with python-telegram-bot)
pytz==2024.1                  # Timezone handling (IST conversions)

# Testing
pytest==8.0.0                 # Testing framework
pytest-asyncio==0.23.0        # Async test support
pytest-cov==4.1.0             # Code coverage reports
